Revert "Backport/pim 2747"
