+++ spec/Pim/Bundle/BaseConnectorBundle/Reader/Doctrine/ProductReaderSpec.php
@@ -2,7 +2,6 @@
 
 namespace spec\Pim\Bundle\BaseConnectorBundle\Reader\Doctrine;
 
-use Doctrine\ORM\EntityManager;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Doctrine\ORM\AbstractQuery;
@@ -14,19 +13,17 @@ use Pim\Bundle\CatalogBundle\Entity\Channel;
 use Doctrine\ORM\QueryBuilder;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface;
-use Doctrine\ORM\Query\Expr\From;
 
-class ORMProductReaderSpec extends ObjectBehavior
+class ProductReaderSpec extends ObjectBehavior
 {
     function let(
         ProductRepositoryInterface $repository,
         ChannelManager $channelManager,
         CompletenessManager $completenessManager,
         MetricConverter $metricConverter,
-        StepExecution $stepExecution,
-        EntityManager $entityManager
+        StepExecution $stepExecution
     ) {
-        $this->beConstructedWith($repository, $channelManager, $completenessManager, $metricConverter, $entityManager);
+        $this->beConstructedWith($repository, $channelManager, $completenessManager, $metricConverter);
 
         $this->setStepExecution($stepExecution);
     }
@@ -41,68 +38,43 @@ class ORMProductReaderSpec extends ObjectBehavior
         $channelManager,
         $repository,
         Channel $channel,
-        From $from,
         QueryBuilder $queryBuilder,
         AbstractQuery $query,
         ProductInterface $sku1,
-        ProductInterface $sku2,
-        ProductInterface $sku3
+        ProductInterface $sku2
     ) {
         $channelManager->getChannelByCode('foobar')->willReturn($channel);
         $repository->buildByChannelAndCompleteness($channel)->willReturn($queryBuilder);
-        $queryBuilder->getRootAliases()->willReturn(['root']);
-        $queryBuilder->getDQLPart('from')->willReturn([$from]);
-        $queryBuilder->select('root.id')->willReturn($queryBuilder);
-        $queryBuilder->resetDQLPart('from')->willReturn($queryBuilder);
-        $from->getFrom()->willReturn('from_table');
-        $from->getAlias()->willReturn('alias_table');
-        $queryBuilder->from('from_table', 'alias_table', 'root.id')->willReturn($queryBuilder);
-        $queryBuilder->groupBy('root.id')->willReturn($queryBuilder);
-
         $queryBuilder->getQuery()->willReturn($query);
-
-        $query->getArrayResult()->willReturn(array_flip([1, 33, 789]));
-
-        $repository->findByIds([1, 33, 789])->willReturn([$sku1, $sku2, $sku3]);
+        $query->execute()->willReturn(array($sku1, $sku2));
 
         $this->setChannel('foobar');
         $this->read()->shouldReturn($sku1);
+        $this->read()->shouldReturn($sku2);
+        $this->read()->shouldReturn(null);
     }
 
     function it_generates_channel_completenesses_first_time_it_reads(
         $channelManager,
         $completenessManager,
         $repository,
-        From $from,
         Channel $channel,
         QueryBuilder $queryBuilder,
         AbstractQuery $query,
         ProductInterface $sku1,
-        ProductInterface $sku2,
-        ProductInterface $sku3
+        ProductInterface $sku2
     ) {
         $channelManager->getChannelByCode('foobar')->willReturn($channel);
         $repository->buildByChannelAndCompleteness($channel)->willReturn($queryBuilder);
-        $queryBuilder->getRootAliases()->willReturn(['root']);
-        $queryBuilder->getDQLPart('from')->willReturn([$from]);
-        $queryBuilder->select('root.id')->willReturn($queryBuilder);
-        $queryBuilder->resetDQLPart('from')->willReturn($queryBuilder);
-        $from->getFrom()->willReturn('from_table');
-        $from->getAlias()->willReturn('alias_table');
-        $queryBuilder->from('from_table', 'alias_table', 'root.id')->willReturn($queryBuilder);
-        $queryBuilder->groupBy('root.id')->willReturn($queryBuilder);
-
         $queryBuilder->getQuery()->willReturn($query);
-
-        $query->getArrayResult()->willReturn(array_flip([1, 33, 789]));
-
-        $repository->findByIds([1, 33, 789])->willReturn([$sku1, $sku2, $sku3]);
+        $query->execute()->willReturn(array($sku1, $sku2));
 
         $completenessManager->generateMissingForChannel($channel)->shouldBeCalledTimes(1);
 
         $this->setChannel('foobar');
-        $this->read()->shouldReturn($sku1);
-
+        $this->read();
+        $this->read();
+        $this->read();
     }
 
     function it_converts_metric_values(
@@ -110,75 +82,54 @@ class ORMProductReaderSpec extends ObjectBehavior
         $repository,
         $metricConverter,
         Channel $channel,
-        From $from,
         QueryBuilder $queryBuilder,
         AbstractQuery $query,
         ProductInterface $sku1,
-        ProductInterface $sku2,
-        ProductInterface $sku3
+        ProductInterface $sku2
     ) {
         $channelManager->getChannelByCode('foobar')->willReturn($channel);
         $repository->buildByChannelAndCompleteness($channel)->willReturn($queryBuilder);
-        $queryBuilder->getRootAliases()->willReturn(['root']);
-        $queryBuilder->getDQLPart('from')->willReturn([$from]);
-        $queryBuilder->select('root.id')->willReturn($queryBuilder);
-        $queryBuilder->resetDQLPart('from')->willReturn($queryBuilder);
-        $from->getFrom()->willReturn('from_table');
-        $from->getAlias()->willReturn('alias_table');
-        $queryBuilder->from('from_table', 'alias_table', 'root.id')->willReturn($queryBuilder);
-        $queryBuilder->groupBy('root.id')->willReturn($queryBuilder);
-
         $queryBuilder->getQuery()->willReturn($query);
-
-        $query->getArrayResult()->willReturn(array_flip([1, 33, 789]));
-
-        $repository->findByIds([1, 33, 789])->willReturn([$sku1, $sku2, $sku3]);
-        $this->setChannel('foobar');
+        $query->execute()->willReturn(array($sku1, $sku2));
 
         $metricConverter->convert($sku1, $channel)->shouldBeCalled();
+        $metricConverter->convert($sku2, $channel)->shouldBeCalled();
 
-        $this->read()->shouldReturn($sku1);
+        $this->setChannel('foobar');
+        $this->read();
+        $this->read();
+        $this->read();
     }
 
     function it_increments_read_count_each_time_it_reads(
         $channelManager,
         $repository,
         $stepExecution,
-        $channelManager,
-        $repository,
         Channel $channel,
-        From $from,
         QueryBuilder $queryBuilder,
         AbstractQuery $query,
         ProductInterface $sku1,
-        ProductInterface $sku2,
-        ProductInterface $sku3
+        ProductInterface $sku2
     ) {
         $channelManager->getChannelByCode('foobar')->willReturn($channel);
         $repository->buildByChannelAndCompleteness($channel)->willReturn($queryBuilder);
-        $queryBuilder->getRootAliases()->willReturn(['root']);
-        $queryBuilder->getDQLPart('from')->willReturn([$from]);
-        $queryBuilder->select('root.id')->willReturn($queryBuilder);
-        $queryBuilder->resetDQLPart('from')->willReturn($queryBuilder);
-        $from->getFrom()->willReturn('from_table');
-        $from->getAlias()->willReturn('alias_table');
-        $queryBuilder->from('from_table', 'alias_table', 'root.id')->willReturn($queryBuilder);
-        $queryBuilder->groupBy('root.id')->willReturn($queryBuilder);
-
         $queryBuilder->getQuery()->willReturn($query);
+        $query->execute()->willReturn(array($sku1, $sku2));
 
-        $query->getArrayResult()->willReturn(array_flip([1, 33, 789]));
-
-        $repository->findByIds([1, 33, 789])->willReturn([$sku1, $sku2, $sku3]);
-        $this->setChannel('foobar');
-
-        $stepExecution->incrementSummaryInfo('read')->shouldBeCalledTimes(3);
+        $stepExecution->incrementSummaryInfo('read')->shouldBeCalledTimes(2);
 
         $this->setChannel('foobar');
         $this->read();
         $this->read();
         $this->read();
-        $this->read();
+    }
+
+    function its_read_method_throws_exception_if_channel_cannot_be_found($channelManager)
+    {
+        $channelManager->getChannelByCode('mobile')->willReturn(null);
+
+        $this->setChannel('mobile');
+        $this->shouldThrow(new \InvalidArgumentException('Could not find the channel "mobile"'))->duringRead();
     }
 
     function it_exposes_the_channel_field($channelManager)
