+++ src/Pim/Bundle/BaseConnectorBundle/Reader/Doctrine/BulkProductReader.php
@@ -9,7 +9,7 @@ namespace Pim\Bundle\BaseConnectorBundle\Reader\Doctrine;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BulkProductReader extends ObsoleteProductReader
+class BulkProductReader extends ProductReader
 {
     /**
      * {@inheritdoc}
