--- src/Pim/Bundle/BaseConnectorBundle/Reader/Doctrine/ODMProductReader.php
@@ -1,198 +1,0 @@
-<?php
-
-namespace Pim\Bundle\BaseConnectorBundle\Reader\Doctrine;
-
-use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
-use Doctrine\ODM\MongoDB\DocumentManager;
-use Pim\Bundle\BaseConnectorBundle\Reader\ProductReaderInterface;
-use Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface;
-use Symfony\Component\Validator\Constraints as Assert;
-use Pim\Bundle\TransformBundle\Converter\MetricConverter;
-use Pim\Bundle\BaseConnectorBundle\Validator\Constraints\Channel as ChannelConstraint;
-use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
-use Pim\Bundle\CatalogBundle\Manager\CompletenessManager;
-
-/**
- * Reads products for Mongodb
- *
- * @author    Olivier Soulet <olivier.soulet@akeneo.com>
- * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class ODMProductReader extends AbstractConfigurableStepElement implements ProductReaderInterface
-{
-    /**
-     * @var string
-     *
-     * @Assert\NotBlank(groups={"Execution"})
-     * @ChannelConstraint
-     */
-    protected $channel;
-
-    /**
-     * @var StepExecution
-     */
-    protected $stepExecution;
-
-    /**
-     * @var ChannelManager
-     */
-    protected $channelManager;
-
-    /**
-     * @var AbstractQuery
-     */
-    protected $query;
-
-    /**
-     * @var Cursor
-     */
-    protected $products;
-
-    /**
-     * @var DocumentManager
-     */
-    protected $documentManager;
-
-    /**
-     * @var ProductRepositoryInterface
-     */
-    protected $repository;
-
-    /**
-     * @var CompletenessManager
-     */
-    protected $completenessManager;
-
-    /**
-     * @var MetricConverter
-     */
-    protected $metricConverter;
-
-    /**
-     * @var boolean
-     */
-    protected $executed = false;
-
-    /**
-     * @param ProductRepositoryInterface $repository
-     * @param ChannelManager             $channelManager
-     * @param CompletenessManager        $completenessManager
-     * @param MetricConverter            $metricConverter
-     * @param DocumentManager            $documentManager
-     */
-    public function __construct(
-        ProductRepositoryInterface $repository,
-        ChannelManager $channelManager,
-        CompletenessManager $completenessManager,
-        MetricConverter $metricConverter,
-        DocumentManager $documentManager
-    ) {
-        $this->documentManager     = $documentManager;
-        $this->repository          = $repository;
-        $this->channelManager      = $channelManager;
-        $this->completenessManager = $completenessManager;
-        $this->metricConverter     = $metricConverter;
-    }
-
-    /**
-     * Get query to execute
-     *
-     * @return Doctrine\ORM\AbstractQuery
-     */
-    public function getQuery()
-    {
-        return $this->query;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function read()
-    {
-        $this->documentManager->clear();
-
-        if (!$this->executed) {
-            $this->executed = true;
-            if (!is_object($this->channel)) {
-                $this->channel = $this->channelManager->getChannelByCode($this->channel);
-            }
-
-            $this->completenessManager->generateMissingForChannel($this->channel);
-
-            $this->query = $this->repository
-                ->buildByChannelAndCompleteness($this->channel)
-                ->getQuery();
-
-            $this->products = $this->getQuery()->execute();
-
-            // MongoDB Cursor are not positioned on first element (whereas ArrayIterator is)
-            // as long as getNext() hasn't be called
-            $this->products->getNext();
-        }
-
-        $result = $this->products->current();
-
-        if ($result) {
-            $this->metricConverter->convert($result, $this->channel);
-            $this->stepExecution->incrementSummaryInfo('read');
-            $this->products->next();
-        }
-
-        return $result;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getConfigurationFields()
-    {
-        return [
-            'channel' => [
-                'type'    => 'choice',
-                'options' => [
-                    'choices'  => $this->channelManager->getChannelChoices(),
-                    'required' => true,
-                    'select2'  => true,
-                    'label'    => 'pim_base_connector.export.channel.label',
-                    'help'     => 'pim_base_connector.export.channel.help'
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setChannel($channel)
-    {
-        $this->channel = $channel;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChannel()
-    {
-        return $this->channel;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setStepExecution(StepExecution $stepExecution)
-    {
-        $this->stepExecution = $stepExecution;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function initialize()
-    {
-        $this->query = null;
-        $this->documentManager->clear();
-        $this->executed = false;
-    }
-}
