+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/processors.yml
@@ -7,7 +7,6 @@ parameters:
     pim_base_connector.processor.product.class:                      Pim\Bundle\BaseConnectorBundle\Processor\ProductProcessor
     pim_base_connector.processor.category.class:                     Pim\Bundle\BaseConnectorBundle\Processor\CategoryProcessor
 
-    pim_base_connector.processor.product_to_flat_array.class:        Pim\Bundle\BaseConnectorBundle\Processor\ProductToFlatArrayProcessor
     pim_base_connector.processor.dummy.class:                        Pim\Bundle\BaseConnectorBundle\Processor\DummyProcessor
 
 services:
@@ -29,12 +28,6 @@ services:
         class: %pim_base_connector.processor.csv_serializer.product.class%
         parent: pim_base_connector.processor.csv_serializer
         arguments:
-            - '@pim_catalog.manager.channel'
-
-    pim_base_connector.processor.product_to_flat_array:
-        class: %pim_base_connector.processor.product_to_flat_array.class%
-        arguments:
-            - '@pim_serializer'
             - '@pim_catalog.manager.channel'
 
     pim_base_connector.processor.dummy:
