+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/readers.yml
@@ -1,4 +1,5 @@
 parameters:
+    pim_base_connector.reader.doctrine.product.class:      Pim\Bundle\BaseConnectorBundle\Reader\Doctrine\ProductReader
     pim_base_connector.reader.doctrine.bulk_product.class: Pim\Bundle\BaseConnectorBundle\Reader\Doctrine\BulkProductReader
     pim_base_connector.reader.orm.entity.class:            Pim\Bundle\BaseConnectorBundle\Reader\ORM\EntityReader
 
@@ -19,7 +20,6 @@ services:
             - '@pim_catalog.manager.channel'
             - '@pim_catalog.manager.completeness'
             - '@pim_transform.converter.metric'
-            - '@pim_catalog.object_manager.product'
 
     pim_base_connector.reader.doctrine.bulk_product:
         parent: pim_base_connector.reader.doctrine.product
