--- src/Pim/Bundle/BaseConnectorBundle/Resources/config/storage_driver/doctrine/orm.yml
@@ -1,2 +1,0 @@
-parameters:
-    pim_base_connector.reader.doctrine.product.class: Pim\Bundle\BaseConnectorBundle\Reader\Doctrine\ORMProductReader
