+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/writers.yml
@@ -7,7 +7,6 @@ parameters:
     pim_base_connector.writer.file.product.class: Pim\Bundle\BaseConnectorBundle\Writer\File\ProductWriter
 
     pim_base_connector.writer.dummy.class: Pim\Bundle\BaseConnectorBundle\Writer\DummyWriter
-    pim_base_connector.writer.file.csv.class: Pim\Bundle\BaseConnectorBundle\Writer\File\CsvWriter
 
 services:
     pim_base_connector.writer.file:
@@ -21,11 +20,6 @@ services:
 
     pim_base_connector.writer.dummy:
         class: %pim_base_connector.writer.dummy.class%
-
-    pim_base_connector.writer.file.csv:
-        class: %pim_base_connector.writer.file.csv.class%
-        arguments:
-            - '@pim_catalog.manager.media'
 
     pim_base_connector.writer.doctrine.product:
         class: %pim_base_connector.writer.doctrine.product.class%
